# Optimize base64 prefix test