Instead of base64-encoding each input byte slice decode base64 prefix bits once and test whether input starts with these decoded bits.

```
goos: linux
goarch: amd64
pkg: github.com/AlexanderYastrebov/wireguard-vanity-key
cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
                      │    main     │                HEAD                │
                      │   sec/op    │   sec/op     vs base               │
FindBatchPoint/1024-8   458.4n ± 2%   424.8n ± 1%  -7.32% (p=0.000 n=10)

                      │    main    │              HEAD              │
                      │    B/op    │    B/op     vs base            │
FindBatchPoint/1024-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                      │    main    │              HEAD              │
                      │ allocs/op  │ allocs/op   vs base            │
FindBatchPoint/1024-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```